fixed: round to ruppees instead of paisa

+++ app/src/main/java/com/miassolutions/milkledger/utils/pdf/expensereport/ExpenseReportModels.kt
@@ -3,21 +3,21 @@ package com.miassolutions.milkledger.utils.pdf.expensereport
 data class ExpenseReceiptPdf(
     val dateRange: String,
     val recordList: List<ExpenseItemRecord>,
-    val totalExpenses:Double,
-    val businessExpenses: Double,
-    val personalExpenses : Double,
+    val totalExpenses: Long,
+    val businessExpenses: Long,
+    val personalExpenses : Long,
     val footerNote: String? = null
 )
 
 data class ExpenseItemRecord(
     val date: String,
     val expenseTitle: String,
-    val expenseAmount: Double,
+    val expenseAmount: Long,
     val expenseType: String
 )
 
 data class PdfExpenseSummary(
-    val totalExpenses : Double,
-    val businessExpenses: Double,
-    val personalExpenses : Double
+    val totalExpenses : Long,
+    val businessExpenses: Long,
+    val personalExpenses : Long
 )
